- use locally stored viewpoint variables in the hardware renderer. - move a few variables from SceneDrawer to FRenderViewpoint.
The global r_viewpoint variable is left alone now to always represent the current viewpoint to the play code.
The main reason behind this change is to reduce the amount of global variables being used by the hardware renderer's scene processing code.

+++ src/hwrenderer/scene/hw_skydome.cpp
@@ -289,7 +289,7 @@ void FSkyDomeCreator::SetupMatrices(FMaterial *tex, float x_offset, float y_offs
 		// smaller sky textures must be tiled. We restrict it to 128 sky pixels, though
 		modelMatrix.translate(0.f, -1250.f, 0.f);
 		modelMatrix.scale(1.f, 128 / 230.f, 1.f);
-		yscale = 128 / texh;	// intentionally left as integer.
+		yscale = float(128 / texh);	// intentionally left as integer.
 	}
 	else if (texh < 200)
 	{
